+++ app/src/main/java/com/boardgamegeek/tasks/ResetCollectionTask.java
@@ -3,6 +3,7 @@ package com.boardgamegeek.tasks;
 import android.content.Context;
 
 import com.boardgamegeek.R;
+import com.boardgamegeek.pref.SyncPrefUtils;
 import com.boardgamegeek.pref.SyncPrefs;
 import com.boardgamegeek.service.SyncService;
 
@@ -27,7 +28,7 @@ public class ResetCollectionTask extends ToastingAsyncTask {
 	@Override
 	protected Boolean doInBackground(Void... params) {
 		if (getContext() == null) return false;
-		SyncPrefs.clearCollection(getContext());
+		SyncPrefUtils.clearCollection(SyncPrefs.getPrefs(getContext()));
 		SyncService.sync(getContext(), SyncService.FLAG_SYNC_COLLECTION);
 		return true;
 	}