+++ app/src/main/java/com/boardgamegeek/ui/PlayFragment.java
@@ -2,6 +2,7 @@ package com.boardgamegeek.ui;
 
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
+import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -24,8 +25,10 @@ import com.boardgamegeek.BggApplication;
 import com.boardgamegeek.R;
 import com.boardgamegeek.events.PlayDeletedEvent;
 import com.boardgamegeek.events.PlaySentEvent;
+import com.boardgamegeek.extensions.PreferenceUtils;
 import com.boardgamegeek.extensions.SwipeRefreshLayoutUtils;
 import com.boardgamegeek.extensions.TaskUtils;
+import com.boardgamegeek.extensions.TextViewUtils;
 import com.boardgamegeek.model.Play;
 import com.boardgamegeek.model.Player;
 import com.boardgamegeek.model.builder.PlayBuilder;
@@ -42,8 +45,8 @@ import com.boardgamegeek.util.DialogUtils.OnDiscardListener;
 import com.boardgamegeek.util.ImageUtils;
 import com.boardgamegeek.util.ImageUtils.Callback;
 import com.boardgamegeek.util.NotificationUtils;
-import com.boardgamegeek.util.PreferencesUtils;
 import com.boardgamegeek.util.UIUtils;
+import com.boardgamegeek.util.XmlConverter;
 import com.boardgamegeek.util.fabric.PlayManipulationEvent;
 import com.crashlytics.android.answers.Answers;
 import com.crashlytics.android.answers.ShareEvent;
@@ -62,6 +65,7 @@ import androidx.loader.app.LoaderManager.LoaderCallbacks;
 import androidx.loader.content.CursorLoader;
 import androidx.loader.content.Loader;
 import androidx.palette.graphics.Palette;
+import androidx.preference.PreferenceManager;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout.OnRefreshListener;
 import butterknife.BindView;
@@ -88,6 +92,8 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 	private String imageUrl;
 	private String heroImageUrl;
 	private boolean isRefreshing;
+	private SharedPreferences prefs;
+	private final XmlConverter xmlConverter = new XmlConverter();
 
 	private Unbinder unbinder;
 	private ListView playersView;
@@ -152,6 +158,7 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 			hasBeenNotified = savedInstanceState.getBoolean(KEY_HAS_BEEN_NOTIFIED);
 		}
 		setHasOptionsMenu(true);
+		prefs = PreferenceManager.getDefaultSharedPreferences(requireContext());
 	}
 
 	private void readBundle(@Nullable Bundle bundle) {
@@ -251,13 +258,11 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 	public boolean onOptionsItemSelected(MenuItem item) {
 		switch (item.getItemId()) {
 			case R.id.menu_discard:
-				DialogUtils.createDiscardDialog(getActivity(), R.string.play, false, false, new OnDiscardListener() {
-					public void onDiscard() {
-						play.dirtyTimestamp = 0;
-						play.updateTimestamp = 0;
-						play.deleteTimestamp = 0;
-						save("Discard");
-					}
+				DialogUtils.createDiscardDialog(getActivity(), R.string.play, false, false, () -> {
+					play.dirtyTimestamp = 0;
+					play.updateTimestamp = 0;
+					play.deleteTimestamp = 0;
+					save("Discard");
 				}).show();
 				return true;
 			case R.id.menu_edit:
@@ -272,14 +277,12 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 			case R.id.menu_delete: {
 				DialogUtils.createThemedBuilder(getContext())
 					.setMessage(R.string.are_you_sure_delete_play)
-					.setPositiveButton(R.string.delete, new OnClickListener() {
-						public void onClick(DialogInterface dialog, int id) {
-							if (play.hasStarted()) cancelNotification();
-							play.end(); // this prevents the timer from reappearing
-							play.deleteTimestamp = System.currentTimeMillis();
-							save("Delete");
-							EventBus.getDefault().post(new PlayDeletedEvent());
-						}
+					.setPositiveButton(R.string.delete, (dialog, id) -> {
+						if (play.hasStarted()) cancelNotification();
+						play.end(); // this prevents the timer from reappearing
+						play.deleteTimestamp = System.currentTimeMillis();
+						save("Delete");
+						EventBus.getDefault().post(new PlayDeletedEvent());
 					})
 					.setNegativeButton(R.string.cancel, null)
 					.setCancelable(true)
@@ -318,7 +321,7 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 	@Subscribe(threadMode = ThreadMode.MAIN, sticky = true)
 	public void onEvent(SyncPlaysByGameTask.CompletedEvent event) {
 		if (play != null && event.getGameId() == play.gameId) {
-			if (!TextUtils.isEmpty(event.getErrorMessage()) && PreferencesUtils.getSyncShowErrors(getContext())) {
+			if (!TextUtils.isEmpty(event.getErrorMessage()) && PreferenceUtils.getSyncShowErrors(prefs)) {
 				// TODO: 3/30/17 change to a snackbar (will need to change from a ListFragment)
 				Toast.makeText(getContext(), event.getErrorMessage(), Toast.LENGTH_LONG).show();
 			}
@@ -331,11 +334,8 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 	private void updateRefreshStatus(final boolean value) {
 		isRefreshing = value;
 		if (swipeRefreshLayout != null) {
-			swipeRefreshLayout.post(new Runnable() {
-				@Override
-				public void run() {
-					if (swipeRefreshLayout != null) swipeRefreshLayout.setRefreshing(isRefreshing);
-				}
+			swipeRefreshLayout.post(() -> {
+				if (swipeRefreshLayout != null) swipeRefreshLayout.setRefreshing(isRefreshing);
 			});
 		}
 	}
@@ -462,7 +462,7 @@ public class PlayFragment extends ListFragment implements LoaderCallbacks<Cursor
 		incompleteView.setVisibility(play.incomplete ? View.VISIBLE : View.GONE);
 		noWinStatsView.setVisibility(play.noWinStats ? View.VISIBLE : View.GONE);
 
-		commentsView.setText(play.comments);
+		TextViewUtils.setTextMaybeHtml(commentsView, xmlConverter.toHtml(play.comments));
 		commentsView.setVisibility(TextUtils.isEmpty(play.comments) ? View.GONE : View.VISIBLE);
 		commentsLabel.setVisibility(TextUtils.isEmpty(play.comments) ? View.GONE : View.VISIBLE);