+++ app/src/main/java/com/boardgamegeek/ui/ThreadFragment.java
@@ -1,6 +1,7 @@
 package com.boardgamegeek.ui;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
 import android.view.LayoutInflater;
@@ -13,6 +14,7 @@ import android.widget.TextView;
 
 import com.boardgamegeek.R;
 import com.boardgamegeek.entities.ForumEntity.ForumType;
+import com.boardgamegeek.extensions.PreferenceUtils;
 import com.boardgamegeek.io.Adapter;
 import com.boardgamegeek.io.BggService;
 import com.boardgamegeek.provider.BggContract;
@@ -22,7 +24,6 @@ import com.boardgamegeek.ui.loader.ThreadSafeResponse;
 import com.boardgamegeek.ui.widget.SafeViewTarget;
 import com.boardgamegeek.util.AnimationUtils;
 import com.boardgamegeek.util.HelpUtils;
-import com.boardgamegeek.util.PreferencesUtils;
 import com.github.amlcurran.showcaseview.ShowcaseView;
 import com.github.amlcurran.showcaseview.ShowcaseView.Builder;
 import com.github.amlcurran.showcaseview.targets.Target;
@@ -34,6 +35,7 @@ import androidx.core.widget.ContentLoadingProgressBar;
 import androidx.fragment.app.Fragment;
 import androidx.loader.app.LoaderManager;
 import androidx.loader.content.Loader;
+import androidx.preference.PreferenceManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.recyclerview.widget.RecyclerView.OnScrollListener;
@@ -41,6 +43,8 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import hugo.weaving.DebugLog;
+
+import static com.boardgamegeek.extensions.PreferenceUtils.INVALID_ARTICLE_ID;
 
 public class ThreadFragment extends Fragment implements LoaderManager.LoaderCallbacks<ThreadSafeResponse> {
 	private static final String KEY_FORUM_ID = "FORUM_ID";
@@ -61,7 +65,8 @@ public class ThreadFragment extends Fragment implements LoaderManager.LoaderCall
 	private ForumType objectType;
 	private ShowcaseView showcaseView;
 	private int currentAdapterPosition = 0;
-	private int latestArticleId = PreferencesUtils.INVALID_ARTICLE_ID;
+	private int latestArticleId = INVALID_ARTICLE_ID;
+	private SharedPreferences prefs;
 
 	Unbinder unbinder;
 	@BindView(android.R.id.progress) ContentLoadingProgressBar progressView;
@@ -87,6 +92,7 @@ public class ThreadFragment extends Fragment implements LoaderManager.LoaderCall
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setHasOptionsMenu(true);
+		prefs = PreferenceManager.getDefaultSharedPreferences(requireContext());
 	}
 
 	@Override
@@ -115,15 +121,15 @@ public class ThreadFragment extends Fragment implements LoaderManager.LoaderCall
 		// If this is called in onActivityCreated as recommended, the loader is finished twice
 		LoaderManager.getInstance(this).initLoader(LOADER_ID, null, this);
 
-		latestArticleId = PreferencesUtils.getThreadArticle(getContext(), threadId);
+		latestArticleId = PreferenceUtils.getThreadArticle(prefs, threadId);
 	}
 
 	@Override
 	public void onPause() {
 		super.onPause();
 
-		if (latestArticleId != PreferencesUtils.INVALID_ARTICLE_ID) {
-			PreferencesUtils.putThreadArticle(getContext(), threadId, latestArticleId);
+		if (latestArticleId != INVALID_ARTICLE_ID) {
+			PreferenceUtils.putThreadArticle(prefs, threadId, latestArticleId);
 		}
 	}
 
@@ -143,7 +149,7 @@ public class ThreadFragment extends Fragment implements LoaderManager.LoaderCall
 	public void onPrepareOptionsMenu(Menu menu) {
 		super.onPrepareOptionsMenu(menu);
 		menu.findItem(R.id.menu_scroll_last).setVisible(
-			latestArticleId != PreferencesUtils.INVALID_ARTICLE_ID &&
+			latestArticleId != INVALID_ARTICLE_ID &&
 				adapter != null &&
 				adapter.getItemCount() > 0);
 		menu.findItem(R.id.menu_scroll_bottom).setVisible(adapter != null && adapter.getItemCount() > 0);
@@ -166,7 +172,7 @@ public class ThreadFragment extends Fragment implements LoaderManager.LoaderCall
 	}
 
 	private void scrollToLatestArticle() {
-		if (latestArticleId != PreferencesUtils.INVALID_ARTICLE_ID) {
+		if (latestArticleId != INVALID_ARTICLE_ID) {
 			scrollToPosition(adapter.getPosition(latestArticleId));
 		}
 	}