+++ app/src/main/java/com/boardgamegeek/ui/dialog/PlayTimeFilterDialog.kt
@@ -18,7 +18,7 @@ class PlayTimeFilterDialog : SliderFilterDialog() {
 
     override val supportsSlider = false
 
-    override fun getPositiveData(context: Context, min: Int, max: Int, checkbox: Boolean): CollectionFilterer {
+    override fun getPositiveData(context: Context, min: Int, max: Int, checkbox: Boolean, ignoreRange: Boolean): CollectionFilterer {
         return PlayTimeFilterer(context).apply {
             this.min = min * tickInterval
             this.max = max * tickInterval
@@ -26,9 +26,9 @@ class PlayTimeFilterDialog : SliderFilterDialog() {
         }
     }
 
-    override fun initValues(filter: CollectionFilterer?): SliderFilterDialog.InitialValues {
+    override fun initValues(filter: CollectionFilterer?): InitialValues {
         val f = filter as PlayTimeFilterer?
-        return SliderFilterDialog.InitialValues(
+        return InitialValues(
                 (f?.min ?: (PlayTimeFilterer.lowerBound * tickInterval)) / tickInterval,
                 (f?.max ?: (PlayTimeFilterer.upperBound * tickInterval)) / tickInterval,
                 f?.includeUndefined ?: false
@@ -36,8 +36,7 @@ class PlayTimeFilterDialog : SliderFilterDialog() {
     }
 
     override fun getPinText(context: Context, value: String): String {
-        val time = (value.toIntOrNull() ?: PlayTimeFilterer.lowerBound) * tickInterval
-        return when (time) {
+        return when (val time = (value.toIntOrNull() ?: PlayTimeFilterer.lowerBound) * tickInterval) {
             PlayTimeFilterer.upperBound -> time.asTime().andMore()
             else -> time.asTime()
         }